score added
Score was added and is displayed at the upper right hand corner of the
game screen.

To changed font, color, location, size, etc must be changed in the
index.html page.  for the text string that is displayed that must be
changed in the index.html and the game.js