scrolling map
the map now scrolls.

when changing map sizes or the size of the canvas (view area) then the
screenDims and mapDims containers need to be adjusted to the new values
and everything should work well.
when creating new elements of the game such as monsters or characters or
new items etc.  look at how wulf or the bullet or the walls or the
goodies or frank get the x and y location to draw in the draw method
copy and paste to the new code.  it is a simple this.x -
this.game.screenDims.x or this.y - this.game.screenDims.y
The bounding box also needs to be adjust too.  just replace this.x with
this.boundingbox.left and this.boundingbox.top.